requirements.txt
Flask
Flask-Cache
Flask-Negotiate
gunicorn
futures
requests
urllib3
rdflib==4.2.1
SPARQLWrapper
python-dateutil
pyld
rdflib-jsonld
jsonpath-rw
shortuuid
agora-py
agora-wot
redis-simple-cache
networkx>=2.0